+++ tools/level_editor/level_tree_view.vala
@@ -155,8 +155,6 @@ public class LevelTreeView : Gtk.Box
 
 		this.pack_start(_filter_entry, false, true, 0);
 		this.pack_start(_scrolled_window, true, true, 0);
-
-		this.get_style_context().add_class("level-tree-view");
 	}
 
 	private bool on_button_pressed(Gdk.EventButton ev)