+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Changelog
 * Various fixes and improvements to the Console.
 * The Editor will now check whether the file being edited is deleted from the Project Browser to ask user for confirmation.
 * Added the ability to set in the Preferences the external editors to use when opening Lua and image files.
+* Custom theme improvements.
 
 0.43.0
 ------